Moved VerifierApi into internal package
