+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/VerifierApi.kt
@@ -1,4 +1,4 @@
-package net.corda.nodeapi
+package net.corda.nodeapi.internal
 
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
